requirements.txt
# model-service/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blessed==1.20.0
boto3==1.34.65
botocore==1.34.65
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
contextlib2==21.6.0
datasets==2.19.0
dill==0.3.8
docker==7.0.0
editor==1.6.6
fastapi==0.110.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.3.1
google-pasta==0.2.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.21.4
idna==3.7
importlib-metadata==6.11.0
inquirer==3.2.4
inquirerpy==0.3.4
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
litellm>=1.37.5
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathos==0.3.2
pfzy==0.3.4
platformdirs==4.2.0
pox==0.3.4
ppft==1.7.6.8
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.0.0
pyarrow-hotfix==0.6
pydantic==2.6.4
pydantic-core==2.16.3
pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1
readchar==4.0.6
referencing==0.34.0
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
runs==1.2.2
s3transfer==0.10.1
safetensors==0.4.3
sagemaker==2.218.0
schema==0.7.5
setuptools==69.5.1
six==1.16.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
starlette==0.37.2
tblib==3.0.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.40.1
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
xmod==1.8.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1